Removing the NERD Commenter plugin.
I never really used this module and the only times I did I was
disappointed by the quality. Too much similar commands, no support for
block comments (each line of a block is commented individually) and so
on. Furthermore, I dislike the WTFPL license. Even the FSF do not
recommend this license.
https://www.gnu.org/licenses/license-list.html#WTFPL